market_maker/requirements.txt
ccxt>=4.2.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
asyncio-throttle>=1.0.2
rich>=13.7.0
pydantic>=2.5.0
numpy>=1.26.0
